+++ tests/EnumTest.php
@@ -125,7 +125,7 @@ class EnumTest extends TestCase
         $this->assertEquals($expectedArray, $array);
     }
 
-    public function test_enum_is_macroable()
+    public function test_enum_is_macroable_with_static_methods()
     {
         Enum::macro('toFlippedArray', function () {
             return array_flip(self::toArray());
@@ -133,6 +133,18 @@ class EnumTest extends TestCase
 
         $this->assertTrue(UserType::hasMacro('toFlippedArray'));
         $this->assertEquals(UserType::toFlippedArray(), array_flip(UserType::toArray()));
+    }
+
+    public function test_enum_is_macroable_with_instance_methods()
+    {
+        Enum::macro('macroGetValue', function () {
+            return $this->value;
+        });
+
+        $this->assertTrue(UserType::hasMacro('macroGetValue'));
+
+        $user = new UserType(UserType::Administrator);
+        $this->assertSame(UserType::Administrator, $user->macroGetValue());
     }
 
     public function test_enum_get_instances()