+++ src/EnumServiceProvider.php
@@ -54,7 +54,7 @@ class EnumServiceProvider extends ServiceProvider
             $enum = $parameters[0] ?? null;
 
             return (new EnumKey($enum))->passes($attribute, $value);
-        }, __('laravelEnum::messages.enum_key'));
+        });
 
         $this->app['validator']->extend('enum_value', function ($attribute, $value, $parameters, $validator) {
             $enum = $parameters[0] ?? null;
@@ -68,13 +68,13 @@ class EnumServiceProvider extends ServiceProvider
             $strict = !! json_decode(strtolower($strict));
 
             return (new EnumValue($enum, $strict))->passes($attribute, $value);
-        }, __('laravelEnum::messages.enum_value'));
+        });
 
         $this->app['validator']->extend('enum', function ($attribute, $value, $parameters, $validator) {
             $enum = $parameters[0] ?? null;
 
             return (new Enum($enum))->passes($attribute, $value);
-        }, __('laravelEnum::messages.enum'));
+        });
     }
 
     /**